Inserted casts to set types correctly.
This works around some warnings about integer sizes and makes the casts explicit.
Most of these give no warning when fewer warning flags are turned on.

These need to be eliminated over time.  However, there are a number of locations where
this will not be possible such as calculation of buffer lengths based on pointer
arithmetic that must be put into a uint16_t.

+++ src/protocols/ab/session.h
@@ -77,7 +77,7 @@ struct ab_session_t {
     uint32_t orig_connection_id;
     uint32_t targ_connection_id;
     uint16_t conn_seq_num;
-    uint32_t conn_serial_number;
+    uint16_t conn_serial_number;
 
     int plc_type;
     uint16_t max_payload_size;
